Switch to flake-utils
Now I understand the value of this, it's more than just a genAttrs
call. It generates stuff like packages."${system}".foo and
checks."${system}".foo for you all from a single input set.

Otherwise you have to reuse your "eachSystem" thin for each of
packages/checks/etc which is annoying because you have to repeat the
import of pkgs.