limmat-config: Start setting runtimeInputs explicitly
Idea here is to be able to run test-golden without going via a
devShell, but also without the old hack of trying to run it inside
the devShell itself.

So instead the config will define the dependency packages of its job
and export them and then we'll add them to the devShell.

This is incomplete but just sets out enough to see hte next steps
forrward.