WIP (README): Setting buildInputs in limmat-config.nix
Idea here is to be able to run test-golden without going via a
devShell, but also without the old hack of trying to run it inside
the devShell itself.

So instead the config will define the dependency packages of its job
and export them and then we'll add them to the devShell.

Currently it fails to set OBJTOOL_WERROR and I can't see why, last
time this problem came up it was actually because Kconfig failed due
to a missing dependency.